Move solution logic fully to cent-js
